Changed encryption signature (#561)
* Changed encryption signature

Removed giving the random as input everywere possible. Refactored the
hybrid encryption, to all encryption/decryption related functions and
structures in one file.
