+++ chain-vote/src/committee.rs
@@ -1,7 +1,5 @@
+use crate::encryption::{HybridCiphertext, PublicKey, SecretKey};
 use crate::gang::{GroupElement, Scalar};
-use crate::gargamel::{PublicKey, SecretKey};
-use crate::hybrid;
-use crate::hybrid::SymmetricKey;
 use crate::math::Polynomial;
 use rand_core::{CryptoRng, RngCore};
 
@@ -38,7 +36,7 @@ pub struct MemberState {
     owner_index: usize,
     apubs: Vec<GroupElement>,
     es: Vec<GroupElement>,
-    encrypted: Vec<(hybrid::HybridCiphertext, hybrid::HybridCiphertext)>,
+    encrypted: Vec<(HybridCiphertext, HybridCiphertext)>,
 }
 
 pub type CRS = GroupElement;
@@ -82,19 +80,9 @@ impl MemberState {
                 let share_shek = pshek.evaluate(&idx);
 
                 let pk = &committee_pks[i];
-                let sym_key_shares = SymmetricKey::new(rng);
-                let sym_key_blinders = SymmetricKey::new(rng);
 
-                let rcomm = Scalar::random(rng);
-                let rshek = Scalar::random(rng);
-                let ecomm =
-                    hybrid::hybrid_encrypt(&pk.0, &sym_key_shares, &share_comm.to_bytes(), &rcomm);
-                let eshek = hybrid::hybrid_encrypt(
-                    &pk.0,
-                    &sym_key_blinders,
-                    &share_shek.to_bytes(),
-                    &rshek,
-                );
+                let ecomm = pk.0.hybrid_encrypt(&share_comm.to_bytes(), rng);
+                let eshek = pk.0.hybrid_encrypt(&share_shek.to_bytes(), rng);
 
                 encrypted.push((ecomm, eshek));
             }
