+++ chain-vote/src/gang/ristretto255.rs
@@ -168,42 +168,6 @@ impl From<bool> for Scalar {
     }
 }
 
-macro_rules! lref {
-    ($lty: ident, $class: ident, $rty: ident, $out: ident, $f: ident) => {
-        impl<'a> $class<$rty> for &'a $lty {
-            type Output = $out;
-
-            fn $f(self, other: $rty) -> Self::Output {
-                self.$f(&other)
-            }
-        }
-    };
-}
-
-macro_rules! rref {
-    ($lty: ident, $class: ident, $rty: ident, $out: ident, $f: ident) => {
-        impl<'b> $class<&'b $rty> for $lty {
-            type Output = $out;
-
-            fn $f(self, other: &'b $rty) -> Self::Output {
-                (&self).$f(other)
-            }
-        }
-    };
-}
-
-macro_rules! nref {
-    ($lty: ident, $class: ident, $rty: ident, $out: ident, $f: ident) => {
-        impl $class<$rty> for $lty {
-            type Output = $out;
-
-            fn $f(self, other: $rty) -> Self::Output {
-                (&self).$f(&other)
-            }
-        }
-    };
-}
-
 //////////
 // FE + FE
 //////////
@@ -216,9 +180,7 @@ impl<'a, 'b> Add<&'b Scalar> for &'a Scalar {
     }
 }
 
-lref!(Scalar, Add, Scalar, Scalar, add);
-rref!(Scalar, Add, Scalar, Scalar, add);
-nref!(Scalar, Add, Scalar, Scalar, add);
+std_ops_gen!(Scalar, Add, Scalar, Scalar, add);
 
 //////////
 // FE - FE
@@ -232,9 +194,7 @@ impl<'a, 'b> Sub<&'b Scalar> for &'a Scalar {
     }
 }
 
-lref!(Scalar, Sub, Scalar, Scalar, sub);
-rref!(Scalar, Sub, Scalar, Scalar, sub);
-nref!(Scalar, Sub, Scalar, Scalar, sub);
+std_ops_gen!(Scalar, Sub, Scalar, Scalar, sub);
 
 //////////
 // FE * FE
@@ -248,9 +208,7 @@ impl<'a, 'b> Mul<&'b Scalar> for &'a Scalar {
     }
 }
 
-lref!(Scalar, Mul, Scalar, Scalar, mul);
-rref!(Scalar, Mul, Scalar, Scalar, mul);
-nref!(Scalar, Mul, Scalar, Scalar, mul);
+std_ops_gen!(Scalar, Mul, Scalar, Scalar, mul);
 
 //////////
 // FE * GE
@@ -276,13 +234,9 @@ impl<'a, 'b> Mul<&'b Scalar> for &'a GroupElement {
     }
 }
 
-lref!(Scalar, Mul, GroupElement, GroupElement, mul);
-rref!(Scalar, Mul, GroupElement, GroupElement, mul);
-nref!(Scalar, Mul, GroupElement, GroupElement, mul);
+std_ops_gen!(Scalar, Mul, GroupElement, GroupElement, mul);
 
-lref!(GroupElement, Mul, Scalar, GroupElement, mul);
-rref!(GroupElement, Mul, Scalar, GroupElement, mul);
-nref!(GroupElement, Mul, Scalar, GroupElement, mul);
+std_ops_gen!(GroupElement, Mul, Scalar, GroupElement, mul);
 
 //////////
 // u64 * GE
@@ -326,9 +280,7 @@ impl<'a, 'b> Add<&'b GroupElement> for &'a GroupElement {
     }
 }
 
-lref!(GroupElement, Add, GroupElement, GroupElement, add);
-rref!(GroupElement, Add, GroupElement, GroupElement, add);
-nref!(GroupElement, Add, GroupElement, GroupElement, add);
+std_ops_gen!(GroupElement, Add, GroupElement, GroupElement, add);
 
 //////////
 // GE - GE
@@ -342,9 +294,7 @@ impl<'a, 'b> Sub<&'b GroupElement> for &'a GroupElement {
     }
 }
 
-lref!(GroupElement, Sub, GroupElement, GroupElement, sub);
-rref!(GroupElement, Sub, GroupElement, GroupElement, sub);
-nref!(GroupElement, Sub, GroupElement, GroupElement, sub);
+std_ops_gen!(GroupElement, Sub, GroupElement, GroupElement, sub);
 
 #[cfg(test)]
 mod test {
